+++ src/request-handler/handlers/request-handler-post/request-handler-post.ts
@@ -2,14 +2,14 @@ import { IncomingMessage, ServerResponse } from "node:http";
 import { v6 as uuidv6 } from 'uuid';
 
 
-import checkRequiredFields from "../../utils/check-required-fields";
-import { ENDPOINTS, IPostRequestRequiredProps } from "../request-handler.type";
-import { getParsedJson } from "../../utils/get-parsed-json";
-import { sendError } from "../../utils/send-error";
-import { RESPONSE_CODES } from "../../constants/response-codes.constant";
-import { getUnsupportedEndpointMessage, INCORRECT_JSON } from "../../utils/get-message";
-import { DATABASE } from "../../database/database";
-import { sendResponse } from "../../utils/send-response";
+import checkRequiredFields from "./check-required-fields";
+import { ENDPOINTS, IPostRequestRequiredProps } from "../../request-handler.type";
+import { getParsedJson } from "../../../utils/get-parsed-json";
+import { sendError } from "../../../utils/send-error";
+import { RESPONSE_CODES } from "../../../constants/response-codes.constant";
+import { getUnsupportedEndpointMessage, INCORRECT_JSON } from "../../../utils/get-message";
+import { DATABASE } from "../../../database/database";
+import { sendResponse } from "../../../utils/send-response";
 
 export default function handleRequestPost (
     path: string,
